# Add UserPayload caching during decoding