--- a/TestTools/StreamChatTestTools/Fixtures/JSONs/CurrentUserPayload.json
+++ b/TestTools/StreamChatTestTools/Fixtures/JSONs/CurrentUserPayload.json
@@ -77,7 +77,6 @@
               "online" : true,
               "name" : "Broken Waterfall",
               "image" : "https:\/\/getstream.io\/random_svg\/?id=broken-waterfall-5&amp;name=Broken+waterfall",
-              "secret_note" : "Anaking is Vader!",
           },
           "channel": {
               "id": "B1DFF9C5-E6A6-4BFA-9375-DC5E8C6852FF",