--- a/Tests/StreamChatTests/APIClient/Endpoints/Payloads/CustomDataHashMap_Tests.swift
+++ b/Tests/StreamChatTests/APIClient/Endpoints/Payloads/CustomDataHashMap_Tests.swift
@@ -19,62 +19,72 @@ final class CustomDataHashMap_Tests: XCTestCase {
     }
 
     func test_channelDetailJSONDecodeWithoutAnyCustomData() throws {
-        try! assertEmptyCustomData(ChannelDetailPayload.self, "ChannelPayload")
+        try assertEmptyCustomData(ChannelDetailPayload.self, "ChannelPayload")
     }
     
     func test_channelDetailJSONDecodeWithCustomData() throws {
-        try! assertCustomData(ChannelDetailPayload.self, "ChannelPayloadWithCustom")
+        try assertCustomData(ChannelDetailPayload.self, "ChannelPayloadWithCustom")
     }
 
     func test_messageJSONDecodeWithoutAnyCustomData() throws {
-        try! assertEmptyCustomData(MessagePayload.self, "MessagePayload")
+        try assertEmptyCustomData(MessagePayload.self, "MessagePayload")
     }
     
     func test_messageJSONDecodeWithCustomData() throws {
-        try! assertCustomData(MessagePayload.self, "MessagePayloadWithCustom")
+        try assertCustomData(MessagePayload.self, "MessagePayloadWithCustom")
     }
     
     func test_messageReactionJSONDecodeWithoutAnyCustomData() throws {
-        try! assertEmptyCustomData(MessageReactionPayload.self, "MessageReactionPayload")
+        try assertEmptyCustomData(MessageReactionPayload.self, "MessageReactionPayload")
     }
     
     func test_messageReactionJSONDecodeWithCustomData() throws {
-        try! assertCustomData(MessageReactionPayload.self, "MessageReactionPayloadWithCustom")
+        try assertCustomData(MessageReactionPayload.self, "MessageReactionPayloadWithCustom")
     }
 
     func test_userJSONDecodeWithoutAnyCustomData() throws {
-        try! assertEmptyCustomData(UserPayload.self, "UserPayload")
+        try assertEmptyCustomData(UserPayload.self, "UserPayload")
     }
     
     func test_userJSONDecodeWithCustomData() throws {
-        try! assertCustomData(UserPayload.self, "UserPayloadWithCustom")
+        try assertCustomData(UserPayload.self, "UserPayloadWithCustom")
     }
     
     func test_currentUserJSONDecodeWithoutAnyCustomData() throws {
-        try! assertEmptyCustomData(CurrentUserPayload.self, "CurrentUserPayload")
+        try assertEmptyCustomData(CurrentUserPayload.self, "CurrentUserPayload")
     }
     
     func test_currentUserJSONDecodeWithCustomData() throws {
-        try! assertCustomData(CurrentUserPayload.self, "CurrentUserPayloadWithCustom")
+        try assertCustomData(CurrentUserPayload.self, "CurrentUserPayloadWithCustom")
     }
 }
 
 // MARK: Test helpers
 
 extension CustomDataHashMap_Tests {
-    func assertEmptyCustomData<T>(_ entity: T.Type, _ fileName: String) throws where T: DecodableEntity {
+    func assertEmptyCustomData<T>(
+        _ entity: T.Type,
+        _ fileName: String,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) throws where T: DecodableEntity {
         let jsonData = XCTestCase.mockData(fromJSONFile: fileName)
         let payload = try JSONDecoder.default.decode(entity.self, from: jsonData)
-        XCTAssertEqual(payload.extraData, [:])
+        XCTAssertEqual(payload.extraData, [:], file: file, line: line)
     }
 
-    func assertCustomData<T>(_ entity: T.Type, _ fileName: String) throws where T: DecodableEntity {
+    func assertCustomData<T>(
+        _ entity: T.Type,
+        _ fileName: String,
+        file: StaticString = #filePath,
+        line: UInt = #line
+    ) throws where T: DecodableEntity {
         let jsonData = XCTestCase.mockData(fromJSONFile: fileName)
         let payload = try JSONDecoder.default.decode(entity.self, from: jsonData)
 
-        XCTAssertEqual(payload.extraData["secret_note"], .string("Anakin is Vader!"))
-        XCTAssertEqual(payload.extraData["good_movies_count"], .number(3))
-        XCTAssertEqual(payload.extraData["awesome"], .bool(true))
+        XCTAssertEqual(payload.extraData["secret_note"], .string("Anakin is Vader!"), file: file, line: line)
+        XCTAssertEqual(payload.extraData["good_movies_count"], .number(3), file: file, line: line)
+        XCTAssertEqual(payload.extraData["awesome"], .bool(true), file: file, line: line)
         XCTAssertEqual(payload.extraData["nested_stuff"], .dictionary(
             [
                 "how_many_times": .number(42), "small": .double(0.001),
@@ -84,6 +94,6 @@ extension CustomDataHashMap_Tests {
                     .number(42)
                 ])
             ]
-        ))
+        ), file: file, line: line)
     }
 }