### 🎯 Goal

With #2084 we improved JSON decoding performance, but we can still do better

### 📝 Summary

JSON decoding bottleneck appears to be the repeated User payloads. Each UserPayload has 3 Dates to be decoded, which is the most expensive primitive to decode.

### 🛠 Implementation

We can cache Dates to improve the performance by a lot, but caching UserPayloads improves the performance further and it's more logical, since UserPayload is repeated multiple times, and Dates are repeated because of this.

UserPayload is safe to decode with any JSONDecoder since we use `userInfo` to check if we have caching enabled on that specific decoder.

### 🎨 Showcase

| Before | After |
| ------ | ----- |
|  970ms   |  370ms  |

### 🧪 Manual Testing Notes

You can use the disabled `ChannelListPayload_Tests.test_decode_bigChannelListPayload` test to measure.

### ☑️ Contributor Checklist
- [x] I have signed the [Stream CLA](https://docs.google.com/forms/d/e/1FAIpQLScFKsKkAJI7mhCr7K9rEIOpqIDThrWxuvxnwUq2XkHyG154vQ/viewform) (required)
- [x] Changelog is updated with client-facing changes
- [x] New code is covered by unit tests
- [x] This change follows zero ⚠️ policy (required)
- [x] Comparison screenshots added for visual changes
- [x] Affected documentation updated (docusaurus, tutorial, CMS)