@nuno-vieira 2022-06-27
@b-onc How is the cache invalidated? What happens if the user was updated? I'm a bit concerned this will add some data inconsistencies issues.